requirements.txt
nltk==3.1
numpy==1.10.4
Pillow==3.1.1
git+https://github.com/coagulant/progressbar-python3@c67d947c2d1bead73c9b67d7761e624dc46d30ed
